invoice styles

+++ home-inspection/src/main/java/ca/inspection/home/inspection/entity/Invoice.java
@@ -6,6 +6,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Entity
@@ -20,7 +21,7 @@ public class Invoice {
     private UUID id;
 
     private String type;
-    private Double fee;
+    private BigDecimal fee;
 
     @ManyToOne
     @JoinColumn(name = "booking_id")
